+++ src/modules/computePlans/ComputePlansSlice.ts
@@ -97,35 +97,6 @@ export const listComputePlans = createAsyncThunk<
     }
 );
 
-interface retrieveComputePlansArgs {
-    computePlanKeys: string[];
-}
-export const retrieveComputePlans = createAsyncThunk<
-    ComputePlanT[],
-    retrieveComputePlansArgs,
-    { rejectValue: string }
->(
-    'computePlans/getMultiple',
-    async ({ computePlanKeys }: retrieveComputePlansArgs, thunkAPI) => {
-        const promises = computePlanKeys.map((computePlanKey) =>
-            ComputePlansApi.retrieveComputePlan(computePlanKey, {
-                signal: thunkAPI.signal,
-            })
-        );
-        let responses;
-        try {
-            responses = await Promise.all(promises);
-            return responses.map((response) => response.data);
-        } catch (error) {
-            if (axios.isAxiosError(error)) {
-                return thunkAPI.rejectWithValue(error.response?.data);
-            } else {
-                throw error;
-            }
-        }
-    }
-);
-
 export const retrieveComputePlan = createAsyncThunk<
     ComputePlanT,
     string,
@@ -303,21 +274,6 @@ export const computePlansSlice = createSlice({
                     state.computePlansLoading = false;
                     state.computePlansError = payload || 'Unknown error';
                 }
-            })
-            .addCase(retrieveComputePlans.pending, (state) => {
-                state.computePlans = [];
-                state.computePlansLoading = true;
-                state.computePlansError = '';
-            })
-            .addCase(retrieveComputePlans.fulfilled, (state, { payload }) => {
-                state.computePlans = payload;
-                state.computePlansLoading = false;
-                state.computePlansError = '';
-            })
-            .addCase(retrieveComputePlans.rejected, (state, { payload }) => {
-                state.computePlans = [];
-                state.computePlansLoading = false;
-                state.computePlansError = payload || 'Unknown error';
             })
             .addCase(retrieveComputePlan.pending, (state) => {
                 state.computePlanLoading = true;